@@ -6,7 +6,7 @@
 class Admin::ApplicationController < ApplicationController
   protect_from_forgery
 
-  before_filter :authorize
+  before_action :authorize
 
   layout 'admin'
 
