@@ -10,7 +10,7 @@ class SessionsController < ApplicationController
 
   # Authenticate user and create a new session.
   def create
-    user = User.find_by_email(params[:email])
+    user = User.find_by(email: params[:email])
     if user && user.authenticate(params[:password])
       session[:user_id] = user.id
       redirect_to admin_root_url, notice: "Logged in!"
